% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{Deriv2}
\alias{Deriv2}
\title{Deriv2}
\usage{
Deriv2(expr, var)
}
\arguments{
\item{expr}{expression}

\item{var}{variable}
}
\description{
Deriv2
}
\examples{
myderiv
Deriv(myfun(x, y), c("x", "y"))
Deriv2(myfun(x, y), c("x", "y"))
Deriv(myfun(x, y), c("x", "y"))
}

